refactor: Move toRegex util out of handlers.
