+++ independent-projects/tools/devtools-common/src/main/java/io/quarkus/devtools/utils/Patterns.java
@@ -1,0 +1,65 @@
+package io.quarkus.devtools.utils;
+
+import java.util.regex.Pattern;
+import java.util.regex.PatternSyntaxException;
+
+public class Patterns {
+
+    public static boolean isExpression(String s) {
+        return s == null || s.isEmpty() ? false : s.contains("*") || s.contains("?");
+    }
+
+    public static Pattern toRegex(final String str) {
+        try {
+            String wildcardToRegex = wildcardToRegex(str);
+            if (wildcardToRegex != null && !wildcardToRegex.isEmpty()) {
+                return Pattern.compile(wildcardToRegex);
+            }
+        } catch (PatternSyntaxException e) {
+            //ignore it
+        }
+        return null;
+    }
+
+    private static String wildcardToRegex(String wildcard) {
+        // don't try with file match char in pattern
+        if (!isExpression(wildcard)) {
+            return null;
+        }
+        StringBuffer s = new StringBuffer(wildcard.length());
+        s.append("^.*");
+        for (int i = 0, is = wildcard.length(); i < is; i++) {
+            char c = wildcard.charAt(i);
+            switch (c) {
+                case '*':
+                    s.append(".*");
+                    break;
+                case '?':
+                    s.append(".");
+                    break;
+                case '^': // escape character in cmd.exe
+                    s.append("\\");
+                    break;
+                // escape special regexp-characters
+                case '(':
+                case ')':
+                case '[':
+                case ']':
+                case '$':
+                case '.':
+                case '{':
+                case '}':
+                case '|':
+                case '\\':
+                    s.append("\\");
+                    s.append(c);
+                    break;
+                default:
+                    s.append(c);
+                    break;
+            }
+        }
+        s.append(".*$");
+        return (s.toString());
+    }
+}
